change runtime interface types into new file
