+++ tests/geometryObjects/plane/plane.test.ts
@@ -1,6 +1,6 @@
 import { Plane } from "../../../src/geometryObjects/plane/plane";
 import { ReadonlyValue } from "../../../src/geometryObjects/value/readonlyValue";
-import * as Interfaces from "../../../src/interfaces";
+import * as Interfaces from "../../../src/interfaces/runtimeInterfaces";
 describe("Plane", () => {
     test.concurrent("constructor", () => {
         const plane = new Plane();
